Rename //testing/util to base
