+++ common/BUILD
@@ -29,7 +29,7 @@ cc_test(
     srcs = ["check_test.cpp"],
     deps = [
         ":check",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -54,8 +54,8 @@ cc_test(
     deps = [
         ":enum_base",
         ":enum_base_test_def",
-        "//testing/util:gtest_main",
-        "//testing/util:test_raw_ostream",
+        "//testing/base:gtest_main",
+        "//testing/base:test_raw_ostream",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -75,8 +75,8 @@ cc_test(
     srcs = ["error_test.cpp"],
     deps = [
         ":error",
-        "//testing/util:gtest_main",
-        "//testing/util:test_raw_ostream",
+        "//testing/base:gtest_main",
+        "//testing/base:test_raw_ostream",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -91,7 +91,7 @@ cc_test(
     srcs = ["indirect_value_test.cpp"],
     deps = [
         ":indirect_value",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -106,7 +106,7 @@ cc_test(
     srcs = ["metaprogramming_test.cpp"],
     deps = [
         ":metaprogramming",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -137,7 +137,7 @@ cc_test(
     srcs = ["string_helpers_test.cpp"],
     deps = [
         ":string_helpers",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -157,8 +157,8 @@ cc_test(
     srcs = ["vlog_test.cpp"],
     deps = [
         ":vlog",
-        "//testing/util:gtest_main",
-        "//testing/util:test_raw_ostream",
+        "//testing/base:gtest_main",
+        "//testing/base:test_raw_ostream",
         "@com_google_googletest//:gtest",
     ],
 )
