+++ common/error_test.cpp
@@ -6,7 +6,7 @@
 
 #include <gtest/gtest.h>
 
-#include "testing/util/test_raw_ostream.h"
+#include "testing/base/test_raw_ostream.h"
 
 namespace Carbon::Testing {
 namespace {
