+++ explorer/ast/BUILD
@@ -82,7 +82,7 @@ cc_test(
         ":ast",
         ":ast_test_matchers",
         "//explorer/base:arena",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -94,7 +94,7 @@ cc_test(
         ":ast",
         ":paren_contents",
         "//explorer/base:arena",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -121,7 +121,7 @@ cc_test(
         ":ast",
         ":paren_contents",
         "//explorer/base:arena",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -134,7 +134,7 @@ cc_test(
         ":ast",
         ":paren_contents",
         "//explorer/base:arena",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
