+++ migrate_cpp/cpp_refactoring/BUILD
@@ -62,7 +62,7 @@ cc_test(
     deps = [
         ":fn_inserter",
         ":matcher_test_base",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//clang:tooling",
     ],
@@ -84,7 +84,7 @@ cc_test(
     deps = [
         ":for_range",
         ":matcher_test_base",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//clang:tooling",
     ],
@@ -108,7 +108,7 @@ cc_test(
     deps = [
         ":matcher_test_base",
         ":var_decl",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//clang:tooling",
     ],
