+++ testing/base/test_raw_ostream.h
@@ -2,8 +2,8 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef CARBON_TESTING_UTIL_TEST_RAW_OSTREAM_H_
-#define CARBON_TESTING_UTIL_TEST_RAW_OSTREAM_H_
+#ifndef CARBON_TESTING_BASE_TEST_RAW_OSTREAM_H_
+#define CARBON_TESTING_BASE_TEST_RAW_OSTREAM_H_
 
 #include <gtest/gtest.h>
 
@@ -39,4 +39,4 @@ class TestRawOstream : public llvm::raw_svector_ostream {
 
 }  // namespace Carbon::Testing
 
-#endif  // CARBON_TESTING_UTIL_TEST_RAW_OSTREAM_H_
+#endif  // CARBON_TESTING_BASE_TEST_RAW_OSTREAM_H_
