+++ toolchain/lexer/BUILD
@@ -26,7 +26,7 @@ cc_test(
     srcs = ["token_kind_test.cpp"],
     deps = [
         ":token_kind",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -95,7 +95,7 @@ cc_test(
         ":test_helpers",
         "//common:check",
         "//common:ostream",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "//toolchain/diagnostics:diagnostic_emitter",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
@@ -148,7 +148,7 @@ cc_test(
         ":test_helpers",
         "//common:check",
         "//common:ostream",
-        "//testing/util:gtest_main",
+        "//testing/base:gtest_main",
         "//toolchain/diagnostics:diagnostic_emitter",
         "@com_google_googletest//:gtest",
         "@llvm-project//llvm:Support",
@@ -208,8 +208,8 @@ cc_test(
     deps = [
         ":tokenized_buffer",
         ":tokenized_buffer_test_helpers",
-        "//testing/util:gtest_main",
-        "//testing/util:test_raw_ostream",
+        "//testing/base:gtest_main",
+        "//testing/base:test_raw_ostream",
         "//toolchain/base:yaml_test_helpers",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:mocks",
